+++ src/test/java/ai/labs/eddi/configs/groups/model/AgentGroupConfigurationTest.java
@@ -137,9 +137,10 @@ class AgentGroupConfigurationTest {
 
     @Test
     void phaseType_allValues() {
-        assertEquals(14, PhaseType.values().length);
+        assertEquals(15, PhaseType.values().length);
         assertNotNull(PhaseType.valueOf("PROPOSAL"));
         assertNotNull(PhaseType.valueOf("BARGAIN"));
+        assertNotNull(PhaseType.valueOf("RETRO"));
         assertNotNull(PhaseType.valueOf("OPINION"));
         assertNotNull(PhaseType.valueOf("CRITIQUE"));
         assertNotNull(PhaseType.valueOf("REVISION"));